ezyzip (2) (1)/flask_app_final/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
binarytree==5.1.0
bleach==3.2.1
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
crypto==1.4.1
cryptography==3.4.7
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
djangify==1.0.0
Django==3.1.7
django-crispy-forms==1.11.2
django-import-export==2.5.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
firebase==3.0.1
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jwcrypto==0.8
kiwisolver==1.3.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.1.0
Naked==0.1.31
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.4
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pbr==5.5.1
pickleshare==0.7.5
Pillow==8.0.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.17.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2020.5
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
RapidConnect==0.0.6
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
shellescape==3.8.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
sseclient==0.0.27
statsmodels==0.12.1
stevedore==3.3.0
sympy==1.7
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xlrd==2.0.1
xlwt==1.3.0
